fix storage delete and a lot of improvements

+++ app/src/main/java/br/ufsc/barcodescanner/utils/FileUtils.java
@@ -5,7 +5,7 @@ import java.io.File;
 public class FileUtils {
 
     public static void clearDir(File dir) {
-        if (dir.isDirectory()) {
+        if (dir != null && dir.isDirectory()) {
             String[] children = dir.list();
             for (int i = 0; i < children.length; i++) {
                 new File(dir, children[i]).delete();
